renaming screen methods and moving to service dir
